kubernetes/apps/eu-health-interop-lab/de-registry/requirements.txt
fastapi
uvicorn[standard]
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
